fix: Modify smelting to allow usage of gold bars on furnace to open crafting interface
